Added directory check to pack script

+++ scripts/pack.sh
@@ -1,3 +1,5 @@
-rm -r ./dist
+if [ -d ./dist ]; then
+    rm -r ./dist
+fi
 mkdir -p ./dist/mods/ImmersiveSuicide
 cp -r Contents/mods/ImmersiveSuicide dist/mods
